+++ Models/basic/TextItem.cs
@@ -6,7 +6,7 @@ public partial class TextItem
     {
     }
 
-    public long text_item_id {get; set;}
+    public int text_item_id {get; set;}
     public required string on_page_text_id {get; set;}
 
     public virtual required IList<int> Languages {get; set;}
