+++ Models/manyToMany/EpisodeXQualityAvailable.cs
@@ -5,7 +5,7 @@ namespace NetflixAPI.Models;
 [PrimaryKey(nameof(episode_id), nameof(quality_id))]
 public class EpisodeXQualityAvailable
 {
-    public long episode_id { get; set; }
-    public long quality_id { get; set; }
+    public int episode_id { get; set; }
+    public int quality_id { get; set; }
 
 }
