+++ Models/manyToMany/ProfileXContentTypePreference.cs
@@ -9,7 +9,7 @@ public partial class ProfileXContentTypePreference
     {
     }
 
-    public long profile_id {get; set;}
-    public long content_type_id {get; set;}
+    public int profile_id {get; set;}
+    public int content_type_id {get; set;}
 
 }
